+++ changelog.md
@@ -21,6 +21,7 @@
 - Schema validation to config
 - Indications of the status for each audio process
 - Silence detection
+- Splash screen (basic)
 
 ### Updated